requirements.txt
streamlit
st-cytoscape
pandas
sqlalchemy
psycopg2-binary
python-dotenv
